requirements.txt
asgiref==3.3.4
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-redis==5.0.0
django-redis-cache==3.0.0
djangorestframework==3.12.4
Faker==8.4.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
gprof2dot==2021.2.21
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pydot==1.4.2
pyflakes==2.3.1
Pygments==2.9.0
pygraphviz==1.7
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
